requirements/production.txt
-r base.txt # includes the base.txt requirements file

django-environ==0.4.5